@@ -1,6 +1,6 @@
 module github.com/free5gc/nef
 
-go 1.18
+go 1.21
 
 require (
 	github.com/antihax/optional v1.0.0
